requirements.txt
Arpeggio==1.4
click==6.6
iPOPO==0.6.4
Jinja2==2.8
jsonrpclib-pelix==0.2.7
MarkupSafe==0.23
textX==1.3.1
